deleting separate sidebarNav component
